webapi/webapi-presentation-w3c-tests/presentation/NavigatorPresentation_attributes_basic.html

<!DOCTYPE html>
<!--
Copyright (c) 2015 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-->

<meta charset="utf-8">
<title>NavigatorPresentation basic attributes</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="author" title="Chunyan Wang" href="mailto:chunyanx.wang@intel.com">
<link rel="help" href="http://www.w3.org/TR/2015/WD-presentation-api-20150701/#interface-navigatorpresentation">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
  test(function() {
    assert_true("presentation" in navigator,
      "navigator should implement presentation");

    test(function() {
      assert_true("startSession" in navigator.presentation,
        "navigator.presentation should implement startSession");
      assert_equals(typeof navigator.presentation.startSession, "function",
        "navigator.presentation.startSession type of");
    }, "Test checks that navigator.presentation.startSession exists");

    test(function() {
      assert_true("joinSession" in navigator.presentation,
        "navigator.presentation should implement joinSession");
      assert_equals(typeof navigator.presentation.joinSession, "function",
        "navigator.presentation.joinSession type of");
    }, "Test checks that navigator.presentation.joinSession exists");

    test(function() {
      assert_true("getAvailability" in navigator.presentation,
        "navigator.presentation should implement getAvailability");
      assert_equals(typeof navigator.presentation.getAvailability, "function",
        "navigator.presentation.getAvailability type of");
    }, "Test checks that navigator.presentation.getAvailability exists");

    test(function() {
      assert_true("getSession" in navigator.presentation,
        "navigator.presentation should implement getSession");
      assert_equals(typeof navigator.presentation.getSession, "function",
        "navigator.presentation.getSession type of");
    }, "Test checks that navigator.presentation.getSession exists");

    test(function() {
      assert_true("getSessions" in navigator.presentation,
        "navigator.presentation should implement getSessions");
      assert_equals(typeof navigator.presentation.getSessions, "function",
        "navigator.presentation.getSessions type of");
    }, "Test checks that navigator.presentation.getSessions exists");

    test(function() {
      assert_true("ondefaultsessionstart" in navigator.presentation,
        "navigator.presentation should implement ondefaultsessionstart");
    }, "Test checks that navigator.presentation.ondefaultsessionstart exists");

    test(function() {
      assert_true("onsessionavailable" in navigator.presentation,
        "navigator.presentation should implement onsessionavailable");
    }, "Test checks that navigator.presentation.onsessionavailable exists");
  }, "navigator should implement presentation");
</script>
